Move database interfaces and classes from ObjectStore.OrMapping to ObjectStore.Database.
