+++ src/ObjectStore.SqlClient/CommandBuilder/InsertCommandBuilder.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Text;
 using System.Data.SqlClient;
 using System.Data.Common;
-using ObjectStore.OrMapping;
+using ObjectStore.Database;
 
 namespace ObjectStore.SqlClient
 {
