+++ src/ObjectStore.Sqlite/CommandBuilder/UpdateCommandBuilder.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using Microsoft.Data.Sqlite;
 using System.Data.Common;
-using ObjectStore.OrMapping;
+using ObjectStore.Database;
 
 namespace ObjectStore.Sqlite
 {
