+++ src/ObjectStore.Sqlite/DataBaseProvider.cs
@@ -1,5 +1,4 @@
 ﻿using ObjectStore.Expressions;
-using ObjectStore.OrMapping;
 using System;
 using System.Collections.Generic;
 using System.Data.Common;
@@ -7,6 +6,7 @@ using Microsoft.Data.Sqlite;
 using System.Linq;
 using System.Threading;
 using System.Xml.Linq;
+using ObjectStore.Database;
 
 namespace ObjectStore.Sqlite
 {
