+++ src/ObjectStore/OrMapping/InheritensObjectProvider/WeakCache.cs
@@ -1,4 +1,5 @@
-﻿using ObjectStore.Interfaces;
+﻿using ObjectStore.Database;
+using ObjectStore.Interfaces;
 using System;
 using System.Collections.Generic;
 using System.Linq;
