+++ src/ObjectStore/OrMapping/Mapping/MemberMapping.cs
@@ -1,9 +1,9 @@
 ﻿using System;
 using System.Reflection;
 using System.Collections.Generic;
-using System.Linq;
 using System.Reflection.Emit;
 using ObjectStore.MappingOptions;
+using ObjectStore.Database;
 
 namespace ObjectStore.OrMapping
 {
